Add saveString, getString function.

+++ src/main/java/ktds/fresh/kafkaredisexample/redis/Point.java
@@ -14,17 +14,17 @@ public class Point implements Serializable {
 
     @Id
     private String id;
-    private Long amount;
+    private int amount;
     private LocalDateTime refreshTime;
 
     @Builder
-    public Point(String id, Long amount, LocalDateTime refreshTime) {
+    public Point(String id, int amount, LocalDateTime refreshTime) {
         this.id = id;
         this.amount = amount;
         this.refreshTime = refreshTime;
     }
 
-    public void refresh(long amount, LocalDateTime refreshTime){
+    public void refresh(int amount, LocalDateTime refreshTime){
         if(refreshTime.isAfter(this.refreshTime)){ // 저장된 데이터보다 최신 데이터일 경우
             this.amount = amount;
             this.refreshTime = refreshTime;
